==> R/00_all.qmd <==
---
title: "Group 7 Project"
subtitle: "Final HTML"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

# Project Title

In this project, we investigate a gene expression dataset for patients with bladder cancer. Our data was acquired from the National Center for Biotechnology Information <https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE31684>. It consists of microarray data from 93 bladder cancer patients. Our goal in this project is to investigate the correlation between gene expression and metastasis occurrence. We are also interested in assessing the relevance of smoking when looking at metastasis, and we looked at the regulation of the genes of this population when it comes to metastasis.

## 

```{r}
#| output: false
#| warning: false
#| echo: false
library("quarto")
```

```{r}
#| output: false
#| warning: false
#| echo: false
quarto_render("01_load.qmd")
quarto_render(input = "02_clean.qmd")
quarto_render(input = "03_augment.qmd")
quarto_render(input = "04_describe.qmd")
quarto_render(input = "05_analysis_1.qmd")
quarto_render(input = "06_analysis_2.qmd")

file.rename(from = "01_load.html",
            to = "../results/01_load.html")
file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
file.rename(from = "03_augment.html",
            to = "../results/03_augment.html")
file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
file.rename(from = "05_analysis_1.html",
            to = "../results/05_analysis_1.html")
file.rename(from = "06_analysis_2.html",
            to = "../results/06_analysis_2.html")
file.rename(from = "00_all.html",
            to = "../results/00_all.html")
```

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis_1.qmd >}}

{{< include 06_analysis_2.qmd >}}
