==> R/01_load.qmd <==
---
title: "Group 7 Project"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

## 01_load

```{r}
#| output: false
#| warning: false
#| echo: false
library(tidyverse)
```

#### Creating data folder if not already exists

If the data folder and subfolder \_raw does not already exist, this code will create it.

```{r}

data_dir <- "../data/"
raw_dir <- "../data/_raw/"

if( !dir.exists(data_dir)){
  dir.create(path = data_dir)
}

if( !dir.exists(raw_dir)){
  dir.create(path = raw_dir)
}

```

#### Download metadata

```{r}

metadata_file <- "GSE31684_meta_data.tsv"
metadata_loc <- "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE31nnn/GSE31684/suppl/GSE31684%5Ftable%5Fof%5Fclinical%5Fdetails.txt.gz"

if( !file.exists(str_c(raw_dir, metadata_file)) ){
  download.file(
    url = metadata_loc,
    destfile = str_c(raw_dir, metadata_file))
}
meta_data <- read_tsv(gzfile("../data/_raw/GSE31684_meta_data.tsv"),  show_col_types = FALSE)

meta_data
```

#### Download gene expression data

```{r}
# creating url
gene_url <- "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE31nnn/GSE31684/matrix/GSE31684_series_matrix.txt.gz"

# this is necessary to not get error message about connection buffer
Sys.setenv(VROOM_CONNECTION_SIZE=500000)

# Retrieve the data directly
# skippin first 82 lines because it is just info
gene_exp_file <- read_tsv(file = gene_url,
                          show_col_types = FALSE,
                          skip = 82)

# Write the data to disk
write_tsv(x = gene_exp_file,
          file = "../data/_raw/gene_exp.tsv.gz")

gene_exp_data <- read_tsv(gzfile("../data/_raw/gene_exp.tsv.gz"),
                          show_col_types = FALSE)

gene_exp_data
```

```{r}
# Removed last line:
gene_exp_data <- gene_exp_data |> 
  filter(!str_detect(ID_REF, "^!"))


```

```{r}
# Transpose the data table:
df_long <- gene_exp_data |>
  pivot_longer(cols = -ID_REF, names_to = "GEO", values_to = "Value")

# Transpose the dataframe from long to wide format
df_wide <- df_long |>
  pivot_wider(names_from = ID_REF, values_from = Value)
```

```{r}
# Join metadata and gene expression data based on patient ID:
data_joined = full_join(meta_data, df_wide, by = join_by(GEO))
```

```{r}
write_tsv(data_joined, "../data/01_dat_load.tsv.gz")
write_tsv(head(data_joined), "../data/01_dat_load_head.tsv.gz")
```
