==> R/02_clean.qmd <==
---
title: "Group 7 Project"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

## 02_clean

```{r}
#| output: false
#| warning: false
#| echo: false
library(tidyverse)
```

```{r}
#| warning: false
Sys.setenv(VROOM_CONNECTION_SIZE=5000000)
data <- read_tsv(gzfile("../data/01_dat_load.tsv.gz"), 
                 show_col_types = FALSE)
head(data)
```

#### Clean Metadata

We are creating an overview table of the metadata in the describe doc, and for that we need to clean up the data a bit. We select two attributes relevant to our analysis: Metastasis and Smoking, as well as gene expressions

```{r}
# Cleaning the data: 
data_clean <- data |> 
  relocate(Metastasis, Smoking) |>   
  select(!ID:Cluster)

data_clean
```

#### Writing the data to be used in 04_describe

```{r}
write_tsv(data_clean, "../data/02_dat_clean_wide.tsv.gz")
```

#### Long data format

We would like the data in a long format, where each gene and gene expression has it's own row, for the following augmentation and analysis.

```{r}
# Making the dataset long:
data_clean_long <- data_clean |>

  select(!GEO) |> 
  pivot_longer(cols=-c(Metastasis, Smoking),
               names_to = "gene",
               values_to = "gene_expression")   

data_clean_long
```

We would like to drop the genes with no variability, as they will not provide useful information to our analysis.

```{r}
# Selecting genes with different average gene expression based on the metastatic and non-metastatic cases:
data_clean_long <- data_clean_long |>
  group_by(gene) |>
  filter(sd(gene_expression)!=0)

data_clean_long
```

#### Writing the data

```{r}
write_tsv(data_clean_long, "../data/02_dat_clean.tsv.gz")
write_tsv(head(data_clean_long), "../data/02_dat_clean_head.tsv.gz")
```
