==> R/03_augment.qmd <==
---
title: "Group 7 Project"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

## 03_augmented

```{r}
#| output: false
#| warning: false
#| echo: false
library(tidyverse)
```

```{r}
data_clean <- read_tsv(gzfile("../data/02_dat_clean.tsv.gz"),
                       show_col_types = FALSE)
data_clean
```

#### Creating new columns

```{r}
# We calculate the t-test p value of the gene_expression by comparring metastasis and non-metastasis for each gene:
data_aug <- data_clean |>
  group_by(gene) |>
  mutate(p_value = t.test(gene_expression[Metastasis == 1], 
                          gene_expression[Metastasis == 0])$p.value) |>
  mutate(is_significant = case_when(p_value <= 0.01 ~ "Yes",
                                    p_value > 0.01 ~ "No" ))
```

```{r}
# We are adding the Log2-Fold-Change as an average value for each gene, and a Log2-Fold-Change as a value for each sample with metastasis for each gene.
 data_aug_log <- data_aug |>
   mutate(log2_fold_change_avg = log2(mean(gene_expression[Metastasis == 1]) / 
                                 mean(gene_expression[Metastasis == 0])),
          log2_fold_change_sample = ifelse(
            Metastasis == 0,
            NA,              
            log2(gene_expression / mean(gene_expression[Metastasis == 0]))))

 data_aug_log
```

#### Write the data

```{r}
write_tsv(data_aug_log, "../data/03_dat_aug.tsv.gz")
```
