==> R/04_describe.qmd <==
---
title: "Group 7 Project"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

## 04_describe

Data is available here: [NCBI](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE31684)

The article published based on the data is available here: [PMC](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3569085/)

**The title of the article is:**

*Combination of a novel gene expression signature with a clinical nomogram improves the prediction of survival in high-risk bladder cancer*

A nomogram is a method to graphically depict a statistical prognostic model that generates a probability of a clinical event, such as cancer recurrence or death.

Urothelial carcinoma of the urinary bladder is the fifth most common cancer in the Western World, and represents 3% of cancers diagnosed globally.

#### About the data we use

Cancer gene expression data from 93 patients undergoing radical cystectomy (RC) between 1993 and 2004. Lymph node dissection was performed in 77 patients; no patient has metastatic disease at the time of RC. Metastatic disease means that the cancer has spread to a different part of the body than where it started. Case selection was restricted to those with frozen specimens with measurable volume of malignancy and adequate percentage of tumor.

#### Data load

```{r}
#| echo: false
#| warning: false
#| output: false

library(table1)
library(tidyverse)
```

```{r}
#| warning: false
Sys.setenv(VROOM_CONNECTION_SIZE=5000000)
data_describe <- read_tsv(
  gzfile("../data/02_dat_clean_wide.tsv.gz"),
  show_col_types = FALSE)

data_describe
```

```{r}
data_clean_sd <- read_tsv(gzfile("../data/02_dat_clean.tsv.gz"),
                            show_col_types = FALSE)

data_clean_sd
```

#### Counting number of genes

```{r}
number_of_genes <- data_describe |> 
  select(c(GEO, ends_with("_at"))) |> 
  pivot_longer(cols = -GEO,
               names_to = "Genes",
               values_to = "Gene_expression") |> 
  distinct(Genes) |> 
  count()

number_of_genes
```

We have a total number of 54,675 genes for each patient in the original dataset. Then we also calculate the standard deviation for each gene, and filtered out all the genes with a standard deviation of 0 which means they were neither up-regulated nor down-regulated.

```{r}
number_of_genes_after_sd <- data_clean_sd |> 
  distinct(gene) |> 
  count()

number_of_genes_after_sd
```

After removing these genes, we end up with 48,932 genes.

#### Table visualizing meta data

```{r}
table_data <- data_describe |> 
  mutate(Metastasis = factor(Metastasis)) |> 
  table1(x = formula(~ Smoking + Metastasis), 
         data = _,
         caption = "Overview of number of patients who have metastasis and smoking status")

table_data
```

In the table we can see that the majority of patients are have smoked for at period of time in their lives, though we do not know for how long and how frequently they smoked in that time. Then we have almost an equal amount for current smokers, and patients who have never smoked.

For metastasis, we see that \~60% do not have metastasis and the rest do.

#### Writing a dataset to be used in presentation

```{r}
data_describe |>
  select(c(Smoking, Metastasis)) |> 
  write_tsv("../data/04_dat_desc.tsv.gz")
```
