==> R/06_analysis_2.qmd <==
---
title: "Group 7 Project"
editor: visual
format:
  html:
    embed-resources: true
author: "Sarah Rosenberg Asmussen (s194689),	Mette Bøge Pedersen (s194679),	Caroline Amalie Bastholm Jensen (s213427), Jaime Noguera Piera (s233773), Yassine Turki (s231735)"
---

```{r}
#| output: false
#| warning: false
#| echo: false
library(tidyverse)
library(readr)
```

## 06_analysis_2

```{r}
data <- read_tsv(gzfile("../data/03_dat_aug.tsv.gz"),
                 show_col_types = FALSE)
data
```

#### Description

In this analysis, we want to explore the potential relationship between smoking habits and the altered regulation of gene expression. Each gene in our data set represents variation in gene expression across different patients. Interestingly, many genes exhibit both up-regulation and down-regulation patterns, which vary among individual patients.

We select the top 13 most significant genes from our previous analysis. To further investigate the regulation of gene expression for each patient with metastasis, we calculate a Log2 Fold Change value for each patient. This was calculated by comparing the gene expression in samples with metastasis to the average gene expression across samples without metastasis for each gene. The formula for a specific gene in observation X is:

$$
\text{Log2 Fold Change}(X) = \log_2\left(\frac{\text{gene expression}_{\text{metastasis}}(X)}{\text{mean}(\text{gene expression}_{\text{no-metastasis}})}\right)
$$

This figure shows the top 13 most significant genes, and their average Log2 Fold Change value (circle), together with their lower and upper 95% confidence intervals (bars).

**Conclusion of second analysis:**

From this figure, we can see that 5 genes are slightly up-regulated, and the rest 8 genes are down-regulated. For the gene 206561_s_at, being the most down-regulated, we also observe larger variability across these samples which is reflected in a larger error-bar.

```{r}
conf_int_plot <- data |>
  group_by(gene) |>
  filter(Metastasis == 1) |>
  filter(p_value < 0.001) |>
  ggplot(aes(x = log2_fold_change_sample,
             y = reorder(gene, 
                         log2_fold_change_sample, 
                         FUN = mean))) + 
  stat_summary(fun.data = mean_cl_normal, 
               geom = "point", 
               shape = 20, 
               size = 2, 
               color = "black") +
  stat_summary(fun.data = mean_cl_normal, 
               geom = "errorbar", 
               width = 0.7, 
               color = "black", 
               size = 1) + 
  geom_vline(xintercept = 0, 
             linetype="dotted", 
             color = "blue", 
             size=1) + 
  xlab("Log2 Fold Change (95% CIs)") + 
  ylab("Genes") + 
  ggtitle("Regulation of Genes Associated with Metastasis in Bladder Cancer")

ggsave(
  filename = "../results/06_conf_int_plot.png",
  plot = conf_int_plot,
  device = "png",
  height = 5,
  width = 7.5,
  dpi = 300  
)

print(conf_int_plot)
dev.off()
```

We now want to investigate the proportion of Current Smokers, Former Smokers, and Non-smokers for each gene. Additionally, we want to compare these proportions based on how the gene expression is regulated (whether a gene is up-regulated or down-regulated).

```{r}
smoking_plot <- data |>
  filter(Metastasis == 1) |>
  filter(p_value < 0.001) |>
  mutate(Regulation = case_when(log2_fold_change_sample > 0 ~ "Samples with upregulation",
                                 log2_fold_change_sample < 0 ~ "Samples with downregulation"),
         Avg_Regulation = case_when(log2_fold_change_avg < 0 ~ "Risk if downregulated",
                                    log2_fold_change_avg > 0 ~ "Risk if upregulated")) |>
    ggplot(aes(y = reorder(gene,
                         log2_fold_change_sample, 
                         FUN = mean),
               fill = Smoking)) +
      geom_bar() +
      facet_grid(Avg_Regulation ~ Regulation, scales="free", space = "free") +
      labs(x = "Counts", y = "Genes") +
      ggtitle("Relationship Between Smoking Habits and the Altered Gene Expression") 

ggsave(
  filename = "../results/06_smoking_plot.png",
  plot = smoking_plot,
  device = "png",
  height = 6,
  width = 9.5,
  dpi = 300  
)

print(smoking_plot)
dev.off()
```
